+++ CHANGELOG.md
@@ -41,8 +41,8 @@
   * fixed encoding of parenthesis  
 * added shorthand grunt tasks `grunt check-code` and `grunt fix-code` for linting js and scss
 * added npm command `npm run generator` to run latest generator-nikita locally
-* added optional libraries `siema`, `choices`, `react-select`, `a11y-dialog`, `lodash` and `date-fns` 
-* removed optional libraries `jquery`, `swiper` and `select2`
+* added optional libraries `choices`, `react-select`, `a11y-dialog`, `lodash` and `date-fns` 
+* removed optional libraries `jquery`, and `select2`
 * removed `uglifyJS`
 * removed `assemble`
 * removed `grunt-accessibility`